Add size changed handlers once
If the control is reused, i.e. when navigation view pages are cached, new size changed event handlers were added each time the control was loaded. That could result in multiple size change handler invocations. Instead of adding the event handlers on load, add once in OnApplyTemplate() and check IsLoaded in the handler.