Fix build break when consumer targets WAS 1.8.0 preview
When the library was built using Win App Sdk 1.6 extra entries were added to the library's pri file in error. They are for the webview2 loader, one for each native platform. This didn't cause any issues until the nuget was added to a project targeting Win App Sdk 1.8 preview. That caused the build to attempt to copy the loader to the output, but as they didn't exist the build was broken.